requirements.txt
flask>=3.0.0
yt-dlp>=2024.1.1
Pillow>=10.0.0
pystray>=0.19.0
pyngrok>=7.0.0
